+++ CHANGES.md
@@ -1,4 +1,5 @@
 #### [unreleased]
+* added check to see if wp-cron is updating and if not revert to previous updating
 
 #### 7.4.1 / 2017-10-22
 * oops, during refactor of `Install` I copied the incorrect query for GitHub's remote install